% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAtools-package.R
\docType{data}
\name{dbExample}
\alias{dbExample}
\title{Simulated database with 1,000 individuals}
\format{
A data frame with each row being a DNA profile and each column a
part of a genetic marker. Note that homozygote profiles has the same allelic
value in the two columns associated to the same marker.
}
\description{
Database containing 1,000 simulated DNA profiles typed on ten autosomal
markers.
}
