% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqEst.R
\name{freqEst}
\alias{freqEst}
\title{Simple allele frequency estimation}
\usage{
freqEst(x)
}
\arguments{
\item{x}{A database of the form ['id','locus1 allele1','locus1
allele2',...,'locusN allele 1','locusN allele2'].}
}
\value{
Returns a list of probability vectors - one vector for each locus.
}
\description{
Estimates allele frequencies from a database with DNA profiles
}
\details{
Computes the allele frequencies for a given database.
}
\examples{

  data(dbExample)
  freqEst(dbExample)

}
\author{
James Curran and Torben Tvedebrink
}
