% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{genRypeRec}
\alias{genRypeRec}
\title{Generates DNA profiles of n individuals.}
\usage{
genRypeRec(x, t, k, n, print = FALSE)
}
\arguments{
\item{x}{Allele probabilities}

\item{t}{theta correction}

\item{k}{Relatedness vector}

\item{n}{Number of probles}

\item{print}{Print information}
}
\description{
These are formed as n/2 pairs for relatives with
a IDB-vector given by k. I.e. the profiles are 
mutually unrelated between pairs.
}
