% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{genTypeRec}
\alias{genTypeRec}
\title{Generates DNA profiles of n unrelated individuals for a locus}
\usage{
genTypeRec(x, t, n, z = rep(0, lx <- length(x)))
}
\arguments{
\item{x}{Allele probabilities}

\item{t}{theta correction}

\item{n}{Number of probles}

\item{z}{FIXME}
}
\description{
Generates DNA profiles of n unrelated individuals for a locus
}
