% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcompare.R
\name{print.dbcompare}
\alias{print.dbcompare}
\title{Prints the summary matrix}
\usage{
\method{print}{dbcompare}(x, ...)
}
\arguments{
\item{x}{Summary matrix returned from dbcompare}

\item{...}{...}
}
\value{
Prints the summary matrix and data frame with 'big hits'
}
\description{
Prints the summary matrix and possible 'big hits'.
}
\details{
Prints the summary matrix
}
\examples{

  \dontrun{
  data(dbExample)
  M = dbCompare(dbExample,hit=5)
  M
  }

}
\seealso{
dbCompare,plot.dbcompare
}
\author{
James Curran and Torben Tvedebrink
}
